\name{BETR-package}
\alias{BETR-package}
\alias{BETR}
\docType{package}
\title{
\packageTitle{BETR}
}
\description{
\packageDescription{BETR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BETR}
\packageIndices{BETR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BETR}

Maintainer: \packageMaintainer{BETR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
